requirements.txt
django>=1.5,<1.6
-e git+https://github.com/PaulUithol/backbone-tastypie.git@84acd7b33f73cee98088f5d851e2e6df7443adf6#egg=Backbone_tastypie-dev
django-tastypie==0.9.15
